+++ compiler/src/main/java/info/teksol/mc/mindcode/compiler/ast/nodes/AstFunctionArgument.java
@@ -57,7 +57,7 @@ public class AstFunctionArgument extends AstFragment {
     }
 
     public boolean isInput() {
-        return inModifier || !outModifier;
+        return !refModifier && (inModifier || !outModifier);
     }
 
     public boolean isOutput() {