+++ compiler/src/main/java/info/teksol/mc/mindcode/compiler/ast/nodes/AstLiteral.java
@@ -44,4 +44,8 @@ public abstract class AstLiteral extends AstExpression {
     public abstract double getDoubleValue();
 
     public abstract long getLongValue();
+
+    public int getIntValue() {
+        return (int) getLongValue();
+    }
 }