+++ compiler/src/main/java/info/teksol/mc/mindcode/compiler/evaluator/CompileTimeEvaluatorContext.java
@@ -1,6 +1,7 @@
 package info.teksol.mc.mindcode.compiler.evaluator;
 
 import info.teksol.mc.mindcode.compiler.CompilerContext;
+import info.teksol.mc.mindcode.compiler.callgraph.CallGraph;
 import info.teksol.mc.mindcode.compiler.generation.variables.Variables;
 import info.teksol.mc.mindcode.logic.instructions.InstructionProcessor;
 import info.teksol.mc.profile.CompilerProfile;
@@ -10,5 +11,6 @@ import org.jspecify.annotations.NullMarked;
 public interface CompileTimeEvaluatorContext extends CompilerContext {
     CompilerProfile globalCompilerProfile();
     InstructionProcessor instructionProcessor();
+    CallGraph callGraph();
     Variables variables();
 }